server/requirements.txt
openenv-core>=0.2.3
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
websockets>=12.0
